+++ src/many/src/server/module.rs
@@ -23,7 +23,8 @@ macro_rules! reexport_module {
 reexport_module!(
     base: _0_base;
     blockchain: _1_blockchain;
-    ledger: _2_ledger + _4_ledger_transactions + _6_ledger_commands;
+    ledger: _2_ledger + _6_ledger_commands;
+    events: _4_events;
     kvstore: _3_kvstore + _7_kvstore_commands;
     r#async: _8_async;
     account: _9_account;
